Fixed a few warnings and disabled abort on lint.
The lint error reports if your simply using an older version of a library.
We don't want users builds to stop simply because a newer version of a
library is available, but this is useful during our continuous builds or
development builds, so it's now a project property.

You can build the native plugin alone with: ../gradlew -PlintAbortOnError build
Or the whole project with: ./build_all
which will enable lintAbort automatically.

Change-Id: I96f3b4c7107e200e2cc0548aee39c468d42db914